mock_data/requirements.txt
Faker==36.1.0
python-dotenv==1.0.1
Requests==2.32.3
tenacity==9.0.0
transformers==4.49.0
